+++ src/Observer.php
@@ -11,7 +11,7 @@
  * bundled with this package in the LICENSE file.
  *
  * @package    Platform Menus extension
- * @version    3.0.0
+ * @version    3.1.0
  * @author     Cartalyst LLC
  * @license    Cartalyst PSL
  * @copyright  (c) 2011-2015, Cartalyst LLC
@@ -89,7 +89,7 @@ class Observer
             // will be removed.
             $slugs = [];
             array_walk_recursive($children, function ($value, $key) use (&$slugs) {
-                if ($key == 'slug') {
+                if ($key === 'slug') {
                     $slugs[] = $value;
                 }
             });
@@ -165,7 +165,7 @@ class Observer
             $slugs = [];
             if (is_array($children) and ! empty($children)) {
                 array_walk_recursive($children, function ($value, $key) use (&$slugs) {
-                    if ($key == 'slug') {
+                    if ($key === 'slug') {
                         $slugs[] = $value;
                     }
                 });