+++ tests/MenuModelTest.php
@@ -11,7 +11,7 @@
  * bundled with this package in the LICENSE file.
  *
  * @package    Platform Menus extension
- * @version    3.0.0
+ * @version    3.1.0
  * @author     Cartalyst LLC
  * @license    Cartalyst PSL
  * @copyright  (c) 2011-2015, Cartalyst LLC
@@ -85,21 +85,6 @@ class MenuModelTest extends IlluminateTestCase
         $this->assertEquals(json_encode($roles), array_get($this->menu->getAttributes(), 'roles'));
 
         $this->assertSame($roles, $this->menu->roles);
-    }
-
-    /** @test */
-    public function it_can_retrieve_guarded_attributes()
-    {
-        $guarded = [
-            'lft',
-            'rgt',
-            'menu',
-            'depth',
-            'created_at',
-            'updated_at',
-        ];
-
-        $this->assertSame($guarded, $this->menu->getGuarded());
     }
 
     /** @test */
@@ -182,7 +167,7 @@ class MenuModelTest extends IlluminateTestCase
             ->andReturn($processor = m::mock('Illuminate\Database\Query\Processors\Processor'));
 
         $processor->shouldReceive('processSelect')
-            ->twice()
+            ->times(3)
             ->andReturn([]);
 
         $connection->shouldReceive('select');