+++ tests/MenuRepositoryTest.php
@@ -11,7 +11,7 @@
  * bundled with this package in the LICENSE file.
  *
  * @package    Platform Menus extension
- * @version    3.0.0
+ * @version    3.1.0
  * @author     Cartalyst LLC
  * @license    Cartalyst PSL
  * @copyright  (c) 2011-2015, Cartalyst LLC
@@ -34,6 +34,7 @@ class MenuRepositoryTest extends IlluminateTestCase
 
         // Additional Bindings
         $this->app['sentinel.menus']              = m::mock('Cartalyst\Sentinel\Menus\MenuRepositoryInterface');
+        $this->app['platform.roles']              = m::mock('Platform\Roles\Repositories\RoleRepositoryInterface');
         $this->app['platform.menus.handler.data'] = m::mock('Platform\Menus\Handlers\DataHandlerInterface');
         $this->app['platform.menus.manager']      = m::mock('Platform\Menus\Repositories\ManagerRepository');
         $this->app['platform.menus.validator']    = m::mock('Cartalyst\Support\Validator');
@@ -286,6 +287,10 @@ class MenuRepositoryTest extends IlluminateTestCase
     /** @test */
     public function it_can_retrieve_prepared_menus()
     {
+        $this->app['platform.roles']
+            ->shouldReceive('findAll')
+            ->once();
+
         $model = $this->shouldReceiveCreateModel(2);
 
         $model->shouldReceive('findAll')
@@ -298,18 +303,6 @@ class MenuRepositoryTest extends IlluminateTestCase
                 $callback();
                 return true;
             }))
-            ->andReturn([]);
-
-        $this->app['sentinel']->shouldReceive('getRoleRepository')
-            ->once()
-            ->andReturn($roleRepository = m::mock('Cartalyst\Sentinel\Roles\RoleRepositoryInterface'));
-
-        $roleRepository->shouldReceive('createModel')
-            ->once()
-            ->andReturn($role = m::mock('Cartalyst\Roles\EloquentRole'));
-
-        $role->shouldReceive('all')
-            ->once()
             ->andReturn([]);
 
         $this->app['platform.menus.manager']->shouldReceive('getTypes')
@@ -327,6 +320,10 @@ class MenuRepositoryTest extends IlluminateTestCase
     /** @test */
     public function it_create_a_new_model_if_the_passed_id_cannot_be_found()
     {
+        $this->app['platform.roles']
+            ->shouldReceive('findAll')
+            ->once();
+
         $model = $this->shouldReceiveCreateModel(2);
 
         $model->shouldReceive('findAll')
@@ -339,18 +336,6 @@ class MenuRepositoryTest extends IlluminateTestCase
                 $callback();
                 return true;
             }))
-            ->andReturn([]);
-
-        $this->app['sentinel']->shouldReceive('getRoleRepository')
-            ->once()
-            ->andReturn($roleRepository = m::mock('Cartalyst\Sentinel\Roles\RoleRepositoryInterface'));
-
-        $roleRepository->shouldReceive('createModel')
-            ->once()
-            ->andReturn($role = m::mock('Cartalyst\Roles\EloquentRole'));
-
-        $role->shouldReceive('all')
-            ->once()
             ->andReturn([]);
 
         $this->app['platform.menus.manager']->shouldReceive('getTypes')